backend_python/requirements.txt
uvicorn==0.24.0
fastapi==0.104.1
python-dotenv==1.0.0
pydantic==1.10.12
httpx==0.25.0
gunicorn==21.2.0
openai>=1.0.0
requests==2.31.0
python-multipart==0.0.7
azure-storage-blob==12.18.2
